*** empty log message ***
git-svn-id: svn+ssh://lsc.dc.uba.ar/var/lib/svn/chemhibri/@67 2258c766-92ac-47c0-ac32-b068d4181763

+++ trunk/VersionIFORT/gauss.f
@@ -1,5 +1,5 @@
       FUNCTION GAUSSN()
-      IMPLICIT REAL*8  (A-H,O-Z)
+      IMPLICIT REAL*16  (A-H,O-Z)
       R1=RANF()
       R2=RANF()
       GAUSSN=DSQRT(-2.D0*DLOG(R1))*DCOS(6.2831853D0*R2)
@@ -8,7 +8,7 @@
 
 
       FUNCTION RANF()
-      IMPLICIT REAL*8 (A-H,O-Z)
+      IMPLICIT real*16 (A-H,O-Z)
       PARAMETER (M1=1048576,M2=324565)
       COMMON/YU/X,IDUM
